+++ opentelemetry-sdk/src/opentelemetry/sdk/_shared_internal/__init__.py
@@ -201,16 +201,12 @@ class BatchProcessor(Generic[Telemetry]):
 
     def emit(self, data: Telemetry) -> None:
         if self._shutdown:
-            _logger.info(
-                "Shutdown called, ignoring %s.", self._exporting
-            )
+            _logger.info("Shutdown called, ignoring %s.", self._exporting)
             return
         if self._pid != os.getpid():
             self._bsp_reset_once.do_once(self._at_fork_reinit)
         if len(self._queue) == self._max_queue_size:
-            _logger.warning(
-                "Queue full, dropping %s.", self._exporting
-            )
+            _logger.warning("Queue full, dropping %s.", self._exporting)
         # This will drop a log from the right side if the queue is at _max_queue_size.
         self._queue.appendleft(data)
         if len(self._queue) >= self._max_export_batch_size: